requirements.txt
APScheduler==3.10.4
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
eventlet==0.33.3
Flask==2.3.3
Flask-APScheduler==1.12.4
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
netifaces==0.10.6
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
pytz==2023.3.post1
requests==2.31.0
rich==13.5.3
simple-websocket==0.10.1
six==1.16.0
socketio==0.2.1
tzlocal==5.0.1
urllib3==2.0.4
Werkzeug==2.3.7
wsproto==1.2.0
